Refactor adapter to use event-drivin query style
This commit addresses the issue which was causing a limited set of results to
be returned for queries which returned greater than 200 records. The
`autofetch` option will continue to fetch records from salesforce until
`maxRecords` has been reached. However, this option is only supported when
using the event-drivin style within jsforce.