Add "runLocal" flag to the connection config
This commit adds a "runLocal" flag to the connection config that
removes the dependency on a network connection. It's main purpose
for now is to help with CI runs but in the future could be
extended to use a local cache so developers do not require a
valid Salesforce connection to work.